+++ src/blocks.js
@@ -161,7 +161,7 @@ CostumeIconMorph, SoundIconMorph, SVG_Costume, embedMetadataPNG*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.blocks = '2022-April-22';
+modules.blocks = '2022-August-03';
 
 var SyntaxElementMorph;
 var BlockMorph;
@@ -603,8 +603,7 @@ SyntaxElementMorph.prototype.labelParts = {
     '%msgHat': {
         type: 'input',
         tags: 'read-only static',
-        menu: 'messagesReceivedMenu',
-        react: 'updateEventUpvar'
+        menu: 'messagesReceivedMenu'
     },
     '%msgSend': {
         type: 'input',
@@ -749,7 +748,7 @@ SyntaxElementMorph.prototype.labelParts = {
         menu: 'shadowedVariablesMenu'
     },
 
-    // code mapping (experimental)
+    // code mapping
 
     '%codeKind': {
         type: 'input',
@@ -805,7 +804,30 @@ SyntaxElementMorph.prototype.labelParts = {
             'definition': ['definition'],
             'category': ['category'],
             'custom?': ['custom?'],
-            'global?': ['global?']
+            'global?': ['global?'],
+            'type': ['type'],
+            'scope': ['scope'],
+            'slots': ['slots'],
+            '~' : null,
+            'defaults': ['defaults'],
+            'menus' : ['menus'],
+            'editables' : ['editables']
+        }
+    },
+    '%byob': {
+        type: 'input',
+        tags: 'read-only static',
+        menu: {
+            'label': ['label'],
+            'definition': ['definition'],
+            'category': ['category'],
+            'type': ['type'],
+            'scope': ['scope'],
+            'slots': ['slots'],
+            '~' : null,
+            'defaults': ['defaults'],
+            'menus' : ['menus'],
+            'editables' : ['editables']
         }
     },
 
@@ -1032,17 +1054,17 @@ SyntaxElementMorph.prototype.labelParts = {
     },
     '%send': {
         type: 'multi',
-        slots: '%msgSend',
-        label: 'and send',
+        slots: ['%msgSend', '%s'],
+        label: ['and send', 'with data'],
         tags: 'static',
-        max: 1
+        max: 2
     },
     '%receive': {
         type: 'multi',
-        slots: '%rcv',
-        label: 'to',
+        slots: ['%rcv', '%s'],
+        label: ['to', 'with data'],
         tags: 'static',
-        max: 1
+        max: 2
     },
     '%scriptVars': {
         type: 'multi',
@@ -1373,7 +1395,7 @@ SyntaxElementMorph.prototype.revertToEmptyInput = function (arg) {
                 }
             }
         } else if (this instanceof MultiArgMorph) {
-            deflt = this.labelPart(this.slotSpec);
+            deflt = this.labelPart(this.slotSpecFor(inp));
         } else if (this instanceof ReporterSlotMorph) {
             deflt = this.emptySlot();
         }
@@ -2378,6 +2400,15 @@ SyntaxElementMorph.prototype.showBubble = function (value, exportPic, target) {
         morphToShow.bounds.setWidth(img.width);
         morphToShow.bounds.setHeight(img.height);
         morphToShow.cachedImage = img;
+        morphToShow.version = value.version;
+        morphToShow.step = function () {
+            if (this.version !== value.version) {
+                img = value.image();
+                this.cachedImage = img;
+                this.version = value.version;
+                this.changed();
+            }
+        };
 
         // support blocks to be dragged out of result bubbles:
         morphToShow.isDraggable = true;
@@ -3554,6 +3585,43 @@ BlockMorph.prototype.developersMenu = function () {
         )
     );
     return menu;
+};
+
+BlockMorph.prototype.isChangeableTo = function (type) {
+    // answer whether it's safe to change my type, e.g. from command to
+    // reporter or from global to sprite-local.
+    // valid types "command", "reporter" and "predicate".
+    //
+    // a block is considered "changeable" if
+    // -------------------------------------
+    // * it's a command & the target type isn't also a command & doesn't have a
+    //   next block & is unattached (e.g. the only expression inside a context).
+    //
+    // * it's a reporter or a predicate & the target type is a command & is
+    //   unattached (e.g. the only expression inside a function context).
+    //
+    // * it's a reporter or a predicate & the target type is also a reporter or
+    //   a predicate the type can always be changed
+
+    var typ = this.type();
+    if (typ === type) {return true; }
+    if (typ === 'command' || type === 'command') {
+        return this.isUnattached();
+    }
+    return true;
+};
+
+BlockMorph.prototype.type = function () {
+    // private
+    return this instanceof CommandBlockMorph ? 'command'
+        : (this.isPredicate ? 'predicate' : 'reporter');
+};
+
+BlockMorph.prototype.isUnattached = function () {
+    // private
+    return ((this.nextBlock && !this.nextBlock()) || !this.nextBlock) &&
+        !(this.parent instanceof SyntaxElementMorph) &&
+        !(this.parent instanceof ScriptsMorph);
 };
 
 BlockMorph.prototype.isInheritedVariable = function (shadowedOnly) {
@@ -7413,6 +7481,7 @@ RingMorph.prototype.vanishForSimilar = function () {
         block.selector === 'reportJSFunction' ||
         block.selector === 'reportAttributeOf' ||
         block.selector === 'reportCompiled' ||
+        block.selector === 'reportThisContext' ||
         (block instanceof RingMorph)
     ) {
         this.parent.replaceInput(this, block);
@@ -7456,16 +7525,19 @@ RingMorph.prototype.fixBlockColor = function (nearest, isForced) {
 RingMorph.prototype.userMenu = function () {
     var menu = new MenuMorph(this);
     if (this.parent instanceof MultiArgMorph &&
-            this.parentThatIsA(ScriptsMorph)) {
+        this.parentThatIsA(ScriptsMorph)
+    ) {
         if (!this.parent.maxInputs ||
-                (this.parent.inputs().length < this.parent.maxInputs)) {
+            (this.parent.inputs().length < this.parent.maxInputs)
+        ) {
             menu.addItem(
                 'insert a slot',
                 () => this.parent.insertNewInputBefore(this)
             );
         }
         if (this.isEmptySlot() &&
-                this.parent.inputs().length > this.parent.minInputs) {
+            this.parent.inputs().length > this.parent.minInputs
+        ) {
             menu.addItem(
                 'delete slot',
                 () => this.parent.deleteSlot(this)
@@ -8024,13 +8096,23 @@ ScriptsMorph.prototype.cleanUp = function () {
 
 ScriptsMorph.prototype.exportScriptsPicture = function () {
     var pic = this.scriptsPicture(),
-        ide = this.world().children[0];
+        ide = this.world().children[0],
+        xml = this.scriptsXML();
     if (pic) {
-        ide.saveCanvasAs(
-            pic,
-            (ide.getProjectName() || localize('untitled')) + ' ' +
-                localize('script pic')
+        if (xml) {
+            ide.saveFileAs(
+                embedMetadataPNG(pic, xml),
+                'image/png',
+                (ide.getProjectName() || localize('untitled')) + ' ' +
+                    localize('script pic')
         );
+        } else {
+            ide.saveCanvasAs(
+                pic,
+                (ide.getProjectName() || localize('untitled')) + ' ' +
+                    localize('script pic')
+            );
+        }
     }
 };
 
@@ -8057,6 +8139,35 @@ ScriptsMorph.prototype.scriptsPicture = function () {
         }
     });
     return pic;
+};
+
+ScriptsMorph.prototype.scriptsXML = function () {
+    // private - answer a container (usually sprite) for all scripts
+    var blockEditor = this.parentThatIsA(BlockEditorMorph),
+        ide = this.world().children[0],
+        scripts = this.children.filter(m => m instanceof BlockMorph),
+        target;
+    if (blockEditor) {
+        return ide.blocksLibraryXML(
+            [blockEditor.definition].concat(
+                blockEditor.definition.collectDependencies(
+                    [],
+                    [],
+                    blockEditor.target
+                )
+            ),
+            null,
+            true
+        );
+    }
+    if (scripts.length === 1) {
+        return scripts[0].toXMLString();
+    }
+    target = this.scriptTarget();
+    if (isSnapObject(target)) {
+        return target.toXMLString();
+    }
+    return null;
 };
 
 ScriptsMorph.prototype.addComment = function () {
@@ -8467,6 +8578,16 @@ ScriptsMorph.prototype.selectForEdit = function () {
     return this;
 };
 
+ScriptsMorph.prototype.droppedImage = function (aCanvas, name, embeddedData) {
+    var ide = this.parentThatIsA(IDE_Morph),
+        blockEditor = this.parentThatIsA(BlockEditorMorph);
+    if (!ide && blockEditor) {
+        ide = blockEditor.target.parentThatIsA(IDE_Morph);
+    }
+    if (!ide) {return; }
+    ide.droppedImage(aCanvas, name, embeddedData, 'scripts');
+};
+
 // ScriptsMorph keyboard support
 
 ScriptsMorph.prototype.edit = function (pos) {
@@ -8609,7 +8730,7 @@ ArgMorph.prototype.justDropped = function () {
 // ArgMorph spec extrapolation (for demo purposes)
 
 ArgMorph.prototype.getSpec = function () {
-    return '%s'; // default
+    return this.type === 'list' ? '%l' : '%s'; // default
 };
 
 // ArgMorph menu
@@ -8622,9 +8743,12 @@ ArgMorph.prototype.userMenu = function () {
     }
     menu = sm || new MenuMorph(this);
     if (this.parent instanceof MultiArgMorph &&
-            this.parentThatIsA(ScriptsMorph)) {
+        this.parentThatIsA(ScriptsMorph) &&
+        !(this.parent.slotSpec instanceof Array)
+    ) {
         if (!this.parent.maxInputs ||
-                (this.parent.inputs().length < this.parent.maxInputs)) {
+            (this.parent.inputs().length < this.parent.maxInputs)
+        ) {
             menu.addItem(
                 'insert a slot',
                 () => this.parent.insertNewInputBefore(this)
@@ -10138,7 +10262,10 @@ InputSlotMorph.prototype.messagesMenu = function (searching) {
 };
 
 InputSlotMorph.prototype.messagesReceivedMenu = function (searching) {
-    var dict = {'any message': ['any message']},
+    var dict = {
+            '__shout__go__': ['__shout__go__'],
+            'any message': ['any message']
+        },
         rcvr,
         stage,
         allNames;
@@ -10475,10 +10602,11 @@ InputSlotMorph.prototype.attributesMenu = function (searching) {
             'bottom' : ['bottom']
         };
     }
+    dict['~'] = null;
+    dict.variables = ['variables'];
     if (obj) {
         varNames = obj.variables.names();
         if (varNames.length > 0) {
-            dict['~'] = null;
             varNames.forEach(name =>
                 dict[name] = name
             );
@@ -12473,8 +12601,8 @@ function MultiArgMorph(
 }
 
 MultiArgMorph.prototype.init = function (
-    slotSpec,
-    labelTxt,
+    slotSpec, // string or array of type strings
+    labelTxt, // string or array of prefix labels
     min,
     eSpec,
     arrowColor,
@@ -12492,7 +12620,9 @@ MultiArgMorph.prototype.init = function (
         i;
 
     this.slotSpec = slotSpec || '%s';
-    this.labelText = localize(labelTxt || '');
+    this.labelText = labelTxt instanceof Array ?
+        labelTxt.map(each => localize(each || ''))
+        : localize(labelTxt || '');
     this.infix = infix || '';
     this.collapse = collapse || '';
     this.minInputs = min || 0;
@@ -12511,7 +12641,11 @@ MultiArgMorph.prototype.init = function (
 
     // label text:
     if (this.labelText || (this.slotSpec === '%cs')) {
-        label = this.labelPart(this.labelText);
+        label = this.labelPart(
+            this.labelText instanceof Array ?
+                this.labelText[0]
+                : this.labelText
+        );
         this.add(label);
         label.hide();
     }
@@ -12758,6 +12892,9 @@ MultiArgMorph.prototype.insertNewInputBefore = function (anInput, contents) {
         this.children.splice(idx, 0, newPart);
     }
     newPart.fixLayout();
+    if (this.parent instanceof BlockMorph) {
+        this.parent.fixLabelColor();
+    }
     this.fixLayout();
     return newPart;
 };
@@ -12765,7 +12902,7 @@ MultiArgMorph.prototype.insertNewInputBefore = function (anInput, contents) {
 // MultiArgMorph arity control:
 
 MultiArgMorph.prototype.addInput = function (contents) {
-    var newPart = this.labelPart(this.slotSpec),
+    var newPart = this.labelPart(this.slotSpecFor(this.inputs().length)),
         i, name, idx;
 
     this.addInfix();
@@ -12796,21 +12933,29 @@ MultiArgMorph.prototype.addInput = function (contents) {
             newPart.setContents('#' + idx);
         }
     } else if (this.elementSpec === '%message') {
-        newPart.setContents(localize('message'));
+        newPart.setContents(localize('data'));
     } else if (this.elementSpec === '%keyName') {
         newPart.setContents(localize('key'));
     }
     newPart.parent = this;
     this.children.splice(idx, 0, newPart);
     newPart.fixLayout();
+    if (this.parent instanceof BlockMorph) {
+        this.parent.fixLabelColor();
+    }
     this.fixLayout();
     return newPart;
 };
 
 MultiArgMorph.prototype.addInfix = function () {
-    var infix;
-    if (this.infix === '' || !this.inputs().length) {return; }
-    infix = this.labelPart(this.infix);
+    var infix,
+        label = this.infix ? this.infix
+        : (this.labelText instanceof Array ?
+            this.labelText[this.inputs().length]
+            : '');
+
+    if (label === '' || !this.inputs().length) {return; }
+    infix = this.labelPart(label);
     infix.parent = this;
     this.children.splice(this.children.length - 1, 0, infix);
 };
@@ -12829,7 +12974,9 @@ MultiArgMorph.prototype.removeInput = function () {
             }
         }
     }
-    if (this.infix !== '') {
+    if (this.infix !== '' ||
+        (this.labelText instanceof Array && this.inputs().length)
+    ) {
         if (this.children.length > 1) {
             this.removeChild(this.children[this.children.length - 2]);
         }
@@ -12874,6 +13021,12 @@ MultiArgMorph.prototype.is3ArgRingInHOF = function () {
         }
     }
     return false;
+};
+
+MultiArgMorph.prototype.slotSpecFor = function (index) {
+    return this.slotSpec instanceof Array ?
+        this.slotSpec[index]
+        : this.slotSpec;
 };
 
 // MultiArgMorph events:
@@ -13611,6 +13764,10 @@ ReporterSlotMorph.prototype.isEmptySlot = function () {
 
 ReporterSlotMorph.prototype.fixLayout = function () {
     var contents = this.contents();
+    if (!contents) {
+        contents = this.emptySlot();
+        this.add(contents);
+    }
     this.bounds.setExtent(contents.extent().add(
         this.edge * 2 + this.rfBorder * 2
     ));